build: migrate shell, devtools, app-framework and config to vite
shell and devtools each already owned a `vite.config.ts` for a different job (the
shell bundle, the devtools dev server); those move to `vite.bundle.config.ts` /
`vite.app.config.ts` so `vite.config.ts` means the library build everywhere.

app-framework's and config's `compile-plugin` tasks emitted a `dist/plugin` tree
nothing resolved — their `exports` already name the vite bundles — so they go;
the `dx-plugin` CLI becomes a `plugin-cli` entry of the same build.
