+++ packages/sdk/app-framework/bin/dx-plugin.mjs
@@ -4,4 +4,4 @@
 // Copyright 2026 DXOS.org
 //
 
-void import('../dist/plugin-cli/node-esm/main.mjs');
+void import('../dist/lib/plugin-cli.mjs');
