+++ packages/sdk/app-framework/moon.yml
@@ -10,35 +10,6 @@ tags:
   - typedoc
   - storybook
 tasks:
-  build:
-    deps:
-      - compile-plugin
-      - compile-plugin-cli
-  compile-plugin:
-    command: dx-compile
-    args:
-      - '--entryPoint=src/vite-plugin/index.ts'
-      - '--outputPath=dist/plugin'
-      - '--platform=node'
-    deps:
-      - ^:build
-      - dx-compile:build
-    outputs:
-      - dist/plugin
-  # The `dx-plugin` bin — the plugin-authoring CLI shipped with the package so out-of-repo plugin
-  # authors run the same toolchain; in-repo, the composer-plugin tag's `prebuild` depends on this.
-  compile-plugin-cli:
-    command: dx-compile
-    args:
-      - '--entryPoint=src/plugin-cli/main.ts'
-      - '--outputPath=dist/plugin-cli'
-      - '--platform=node'
-    deps:
-      - dx-compile:build
-    inputs:
-      - src/plugin-cli/**/*
-    outputs:
-      - dist/plugin-cli
   check-module-structure:
     # Verify headless entrypoints resolve under workerd conditions (post-build artifacts
     # must exist — catches stale `./workerd` export paths like the removed browser bundle).
