+++ packages/sdk/config/moon.yml
@@ -6,22 +6,3 @@ tags:
   - ts-vite-build
   - ts-test
   - pack
-tasks:
-  build:
-    deps:
-      - compile-plugin
-  compile-plugin:
-    command: dx-compile
-    args:
-      - '--entryPoint=src/plugin/esbuild-plugin.ts'
-      - '--entryPoint=src/plugin/rollup-plugin.ts'
-      - '--entryPoint=src/plugin/vite-plugin.ts'
-      - '--outputPath=dist/plugin'
-      - '--platform=node'
-    deps:
-      - ^:build
-      - dx-compile:build
-    inputs:
-      - '@group(production)'
-    outputs:
-      - dist/plugin
