--- packages/sdk/shell/vitest.config.ts
@@ -1,16 +1,0 @@
-//
-// Copyright 2025 DXOS.org
-//
-
-import path from 'node:path';
-import { fileURLToPath } from 'node:url';
-
-import { createConfig } from '../../../vitest.base.config';
-
-const dirname = typeof __dirname !== 'undefined' ? __dirname : path.dirname(fileURLToPath(import.meta.url));
-
-export default createConfig({
-  dirname,
-  browser: 'chromium',
-  storybook: true,
-});
